+++ http-server/go.mod
@@ -1,3 +1,8 @@
 module github.com/Fathi122/envoywasmfilter-test/http-server
 
-go 1.14
+go 1.16
+
+require (
+	github.com/sirupsen/logrus v1.8.1
+	github.com/stretchr/testify v1.7.0
+)
